+++ firmware/src/Vector2D.cpp
@@ -1,17 +1,17 @@
 #include "../include/Vector2D.h"
 
-Vector2D::Vector2D(int x, int y)
+Vector2D::Vector2D(float x, float y)
 {
     this->x = x;
     this->y = y;
 }
 
-int Vector2D::getX()
+float Vector2D::getX()
 {
     return x;
 }
 
-int Vector2D::getY()
+float Vector2D::getY()
 {
     return y;
 }
